Refactor trace writer trace id helper
